+++ RogueFrontier/SpaceObject/Stargate.cs
@@ -18,7 +18,7 @@ public class Stargate : ActiveObject {
     public ColoredGlyph tile => new ColoredGlyph(Color.Violet, Color.DarkBlue, '*');
 
     [JsonProperty]
-    public int id { get; private set; }
+    public long id { get; private set; }
     [JsonProperty]
     public System world { get; private set; }
     [JsonProperty]
